api/requirements.txt
# Django and DRF
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-jazzmin==2.6.0

# Database
psycopg2-binary==2.9.9
python-decouple==3.8

# Authentication & Security
PyJWT==2.8.0
python-dotenv==1.0.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1

# Image Processing
Pillow==10.2.0

# Server
gunicorn==21.2.0

# Utils
requests==2.31.0
urllib3==2.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
oauthlib==3.2.2
defusedxml==0.7.1
pycparser==2.21
python3-openid==3.2.0
sqlparse==0.4.4
typing_extensions==4.9.0